Reviewed-on: https://git.friendi.ca/friendica/friendica-addons/pulls/1452

+++ bluesky/bluesky.php
@@ -463,7 +463,7 @@ function bluesky_jot_nets(array &$jotnets_fields)
 
 function bluesky_cron()
 {
-	$last = DI::keyValue()->get('bluesky_last_poll');
+	$last = (int)DI::keyValue()->get('bluesky_last_poll');
 
 	$poll_interval = intval(DI::config()->get('bluesky', 'poll_interval'));
 	if (!$poll_interval) {
